[FIX] web: start typing text in dropdown not showing on hover
When you hover on the "Start Typing" text in autocomplete dropdown,
it will not visible as its CSS overwrite with the rule and change the text color
as white.

With this commit, we have added the css rule for specifically "Start typing"
dropdown item and its looks consistent as other dropdown items.

Task : 2311983

closes odoo/odoo#56233

X-original-commit: a8251584b882270fe2ab25b078387690bc5395c8
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>