[FIX] crm: do not propagate void email / phone of partner to lead
Consider a lead with valid email and/or phone. Some process assign a partner
to it, like mass convert. This partner has no email and/or phone. We should
not void lead values for email and/or phone, as this is information loss.

This is a followup of ae28ba2f055c3749c2ef1952be045b5f105dc6a1 . Test cases are added to try to
avoid future regressions.

Task ID 2311981
PR #54619

X-original-commit: fabd62820c954894ad9a8fa1a34d9a5040f8ea8d