[FW][FIX] website_event_track_(online, session): do not display colorless tags
Colorless tags on tracks are considered as used in backend only for
internal organization. They should not be displayed in frontend
notably

  * not in search dropdowns;
  * not in agenda;
  * not in cards / list views;

If a tag category holds only colorless tags, do not display as a void search
dropdown in frontend, hide it.

Followup of 0a4310926bb51d30a73e62d03f789054bfbcd219

closes odoo/odoo#55445

X-original-commit: abe8733e7553ffe8711b8454adbf0f39906a36b8
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>