[FIX] mail: disable seen notification for 'channel' channels
This commit fully disables seen notification (vv) on channels with
multiple members.

This feature was already disabled on the UI, but the computation
still happened in the background. This was problematic with some
channels with thousands of members, as it took dozens of seconds
to fully compute seen "notification-related" fields during
`mail/init_messaging`.

Part of task-2308312

closes odoo/odoo#55449

X-original-commit: 6f0a1948bce560b4e934cec0f1503d24d5143309
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>