[FIX] website_profile: _prepare_user_value edit tl ok
In courses home page, we use _prepare_user_value to get values from
search strings as render values.

But for example this cause an issue in case of edit_translations, it
has no value but if present is set to `{'edit_translations': True}` in
context.

But _prepare_user_value will set it to context as:

  {'edit_translations': ''}

Which make translation not work on page that used
`_prepare_user_value(kw)` to get values for rendering.

opw-2305816
closes #55444

closes odoo/odoo#55457

X-original-commit: 11c5daae47b50c26ceebe8feb736352b7a27a85b
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>