[FIX] mrp: force default value to 0
Commit fcc902fce28f0b41c730fb5913ae5d1f79a5a9e8 set a default value for
the stock moves initial demand (default_product_uom_qty). The issue is
the stock_move_line models also have a product_uom_qty field. This field
on the line is also st to 1 in the default get. Which is non desired side
effect.
This commit adds another default_product_uom_qty equals to 0 for the
stock move lines.

Task : 2278147