[FIX] mail: hide rating button in chatter
Before this commit, the rating buttons were displayed inside the
chatter an could be used to "rate" from there. Intend is to display
these buttons only on email: the message in chatter should not have
these buttons.

This commit fixes the problem similarly to what has been done in
13.3. Note that this solution is quite poor, because it could remove
content from other messages that inadvertently hit this CSS selector,
but at least it works for all existing rating messages.

task-2282414

closes odoo/odoo#55591

X-original-commit: 056a5fec70bce47c6e4adf3c928a70d1b8daaaa9
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>