[FIX] sale_(coupon): Fix coupon reflection on invoice
The purpose of this task is to reflect the coupon discount perfectly
from sale order to invoice. This issue is already fixed in 13.0 in
commit: bb5d999 but during refactor of coupon it was removed due to
module split.

So in this commit, improve the default invoicing policy of coupon product.
The default invoicing policy should always be by ordered quantity.
Otherwise, if the user changes the default invoicing policy to
"delivered quantity", coupon won't be applied to the invoice.

closes odoo/odoo#65845

Taskid: 2303244
X-original-commit: 6691334b6060e9cfec487c98aacb580fba52cfb0
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>