[FIX] website: handle 404 publisher only for notFound exception
Before this commit we cehck first if user was publisher before to check if the
exception was a 404 NotFound.

What's means that we make this check (1 query sql) most of the time for no
reason, and worst case, when the exception crashes the cursor, this check will
reraise a new exception because isPublisher check is done with a closed cursor.

closes odoo/odoo#55789

X-original-commit: 48f6f986723e68c31181ecceee258377ecd3d96c
Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>