[REF] point_of_sale: simplify media print
The old implementation manually sets 'display' to 'none' to the elements
we don't want to see in the print. This is tedious because we need
to identify every element to be hidden when in fact we know the only
element that we want to print -- the receipt.

So in this commit, we set the 'visibility' of all elements to 'hidden'
and only the element tree of the receipt is set to visible.