[IMP] point_of_sale,pos_*: new receipt screen
Aside from new look, we also move the email option from payment
screen to receipt screen. This allows emailing the receipt to a
customer that is not in the contact list.

We also removed invoicing in the receipt screen during failed invoicing
from payment screen as it is removed in here: 1c3bb1d93fa0921c1004e40dd5f3827353b48487

closes odoo/odoo#55868

Related: odoo/enterprise#12395
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>