[FIX] website_event(_track): use is_view_active to benefits from the cache
Purpose of this commit is to speedup loading of event pages by using
is_view_active helper. It allows to know if a view (template) is active.
This is generally used for options in fontend templates.

Using is_view_active allows to use a cache, speeding up loading of pages
by removing unnecessary queries.

Task ID-2348382

closes odoo/odoo#58780

X-original-commit: bf24c18c44fc45e934524cdeab3d5d8ba50b4def
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>