[FIX] mail: fix activity scheduling dropdown overflow
The dropdown for scheduling the next activity was overflowing inside the
container div.

After this commit, the dropdown is now attached to body element.
In other terms, the pop-in window to schedule an activity in the web kanban
activity widget is now visible (and not displayed anymore under previous line
in the widget tree view).

Task ID:2301262

closes odoo/odoo#56072

X-original-commit: 8684a63c1c66685643486ee81245703c7b0f39b4
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>