[FIX] hr_recruitment: Fix traceback on applicant quick create
Currently, when got to applicant view and quick create the applicant
in the kanban view there is traceback for department_id.
becasue department_id is computed field which depends on job_id
and it was not set as False due to filter on self.

So in this commit, If job_id is not there then department_id
should be set as False.

closes odoo/odoo#56796

Taskid: 2320885
X-original-commit: 73df518d96bf8800da4053ed3bfee8e9b00a5e5e
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>