[FIX] web: do not display nocontent help if empty
Have an action which help field has the value:
```
<p></p>
```
Display an empty list or kanban

Before this commit, the nocontent helper was displayed even though there was
nothing to display.

After this commit, an empty html tag is the same as an empty string,
so the helper is not displayed.

task-2320318

closes odoo/odoo#56883

X-original-commit: 9da4ef9eb42e7256167a4cf6adb7a56885b4f337
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>